% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\docType{data}
\name{animals}
\alias{animals}
\title{Genetic study with beef animals}
\format{An object of class \code{data.frame}}
\source{
Kuehl (2000) Exercise 5.1
}
\usage{
data(animals)
}
\description{
5 sires, each mated to a separate group of dams. \cr
More explanations in: Ex design, 06_Random_Effects, p. 25 \cr
Example for: Random effect Model
}
\examples{
data(animals)
}
\keyword{datasets}
