% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apm.R
\docType{data}
\name{apm}
\alias{apm}
\title{Data on Air Pollution and Mortality}
\format{An object of class \code{data.frame}}
\usage{
data(apm)
}
\description{
General Motors collected data on 59 US Standard Metropolitan
Statistical Areas for a study whether air pollution contributes
to the age-adjusted mortality of the population \cr
More explanations in: regression script p. 41 \cr
Example for: Multiple linear regression
}
\examples{
data(apm)
}
\keyword{datasets}
