% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baby.R
\docType{data}
\name{baby}
\alias{baby}
\title{Survival after Premature Birth}
\format{An object of class \code{data.frame}}
\source{
Hubbard (1986)
}
\usage{
data(baby)
}
\description{
Data of 247 early born babies \cr
More explanations in: regression script p. 137 \cr
Example for: Logistic regression
}
\examples{
data(baby)
}
\keyword{datasets}
