% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheddar.R
\docType{data}
\name{cheddar}
\alias{cheddar}
\title{Cheese quality}
\format{An object of class \code{data.frame}}
\source{
Oehlert (2010) Example 8.6
}
\usage{
data(cheddar)
}
\description{
Cheese acids as response to bacteria strains \cr
More explanations in: Ex design, Script, p. 41 \cr
Example for: Two-way ANOVA, full factorial
}
\examples{
data(cheddar)
}
\keyword{datasets}
