% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheese.data.R
\docType{data}
\name{cheese.data}
\alias{cheese.data}
\title{Cheese Tasting Experiment}
\format{An object of class \code{data.frame}}
\source{
Oehlert, 2010, Example 12.2
}
\usage{
data(cheese.data)
}
\description{
Bitterness as response to background (urban, rural), rater and cheese type \cr
More explanations in: Ex design, 07_Nested_mixed_Part_I, p. 21 \cr
Example for: Mixed effect Model
}
\examples{
data(cheese.data)
}
\references{
Oehlert, 2010, Example 12.2
}
\keyword{datasets}
