% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dish.R
\docType{data}
\name{dish}
\alias{dish}
\title{Test of 9 different detergents}
\format{An object of class \code{data.frame}}
\usage{
data(dish)
}
\description{
Test of 9 different detergents.  There are three basins that are used simultaneously at the same rate with a different detergent in each basin.  Response is number of plates until foam disappears in a basin. Column 1 is session. Column 2 is treatment (kind of detergen).  Column 3 is response (number of dishes). Treatments 1-4 are detergent base 1 with (3, 2, 1, or 0) parts additive.  Treatments 5-8 are detergent base 2 with (3, 2, 1, or 0) parts additive.  Treatment 9 is a control.\cr
}
\details{
For more information see Oehlert (2000) "First Course in Design and Analysis of Experiments".

Data originally from John P. W. M. (1961) "An application of a balanced incomplete block design" Technometrics.
}
\examples{
data(dish)
}
\keyword{datasets}
