% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergoStool.R
\docType{data}
\name{ergoStool}
\alias{ergoStool}
\title{Ergometrics experiment with stool types}
\format{An object of class \code{data.frame}}
\usage{
data(ergoStool)
}
\description{
See help page of \link[nlme]{ergoStool} \cr
More explanations in: Ex design, 07_Nested_mixed_Part_II, p. 3 \cr
Example for: Mixed effect Model
}
\examples{
data(ergoStool)
}
\keyword{datasets}
