% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gala.R
\docType{data}
\name{gala}
\alias{gala}
\title{Species diversity on the Galapagos Islands}
\format{An object of class \code{data.frame}}
\usage{
data(gala)
}
\description{
See help page of \link[faraway]{gala} \cr
More explanations in: regression script p. 148 \cr
Example for: Poission regression
}
\examples{
data(gala)
}
\keyword{datasets}
