% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasses.R
\docType{data}
\name{grasses}
\alias{grasses}
\title{Tropical grass yield}
\format{An object of class \code{data.frame}}
\source{
Oehlert (2010) Problem 8.6
}
\usage{
data(grasses)
}
\description{
yield as a function of cut height (ht),
fertilizer (fert) and cutting interval (int). \cr
More explanations in: Ex design, Script, p. 54 \cr
Example for: Two-way ANOVA, full factorial
}
\examples{
data(grasses)
}
\keyword{datasets}
