% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leinhardt.R
\docType{data}
\name{Leinhardt}
\alias{Leinhardt}
\title{Data on Infant-Mortality}
\format{An object of class \code{data.frame}}
\usage{
data(Leinhardt)
}
\description{
See help page of \link[carData]{Leinhardt} \cr
More explanations in: regression script p. 30 \cr
Example for: Log transformation
}
\examples{
data(Leinhardt)
}
\keyword{datasets}
