% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Machines.R
\docType{data}
\name{Machines}
\alias{Machines}
\title{Productivity Scores for Machines and Workers}
\format{An object of class \code{data.frame}}
\usage{
data(Machines)
}
\description{
See help page of \link[nlme]{Machines} \cr
More explanations in: Ex design, 07_Nested_mixed_Part_II, p. 10 \cr
Example for: Mixed effect Model
}
\examples{
data(Machines)
}
\keyword{datasets}
