% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meat.R
\docType{data}
\name{meat}
\alias{meat}
\title{Study on Meat Storage}
\format{An object of class \code{data.frame}}
\source{
Kuehl (2000), Ex 2.1
}
\usage{
data(meat)
}
\description{
Effect of packaging on bacterial growth of stored meat \cr
More explanations in: Ex design, 02_CRD, p. 2 \cr
Example for: One-way ANOVA
}
\examples{
data(meat)
}
\keyword{datasets}
