% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oats.R
\docType{data}
\name{oats}
\alias{oats}
\title{Data from an Oats Field Trial}
\format{An object of class \code{data.frame}}
\usage{
data(oats)
}
\description{
See help page of \link[MASS]{oats} \cr
More explanations in: Ex design, 08_Split_plots, p. 10 \cr
Example for: Split Plot Design
}
\examples{
data(oats)
}
\keyword{datasets}
