% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pastes.R
\docType{data}
\name{Pastes}
\alias{Pastes}
\title{Paste strength by batch and cask}
\format{An object of class \code{data.frame}}
\usage{
data(Pastes)
}
\description{
See help page of \link[lme4]{Pastes} \cr
More explanations in: Ex design, 07_Nested_mixed_Part_I, p. 12 \cr
Example for: Fully nested Design
}
\examples{
data(Pastes)
}
\keyword{datasets}
