% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pine.R
\docType{data}
\name{pine}
\alias{pine}
\title{Study on pine seedlings needleweight}
\format{An object of class \code{data.frame}}
\source{
Linder, A. and W. Berchtold, (1982)
}
\usage{
data(pine)
}
\description{
Effect of origin and light exposure on needleweight
of 20 three-week old pine seedlings \cr
More explanations in: Ex design, 04_Factorials_Part_I, p. 3 \cr
Example for: Two-way ANOVA, full factorial
}
\examples{
data(pine)
}
\keyword{datasets}
