% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prestige.R
\docType{data}
\name{Prestige}
\alias{Prestige}
\title{Prestige of Canadian Occupations}
\format{An object of class \code{data.frame}}
\usage{
data(Prestige)
}
\description{
See help page of \link[carData]{Prestige} \cr
More explanations in: regression script p. 81 \cr
Example for: Diagnostic plots
}
\examples{
data(Prestige)
}
\keyword{datasets}
