% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snails.R
\docType{data}
\name{snails}
\alias{snails}
\title{Fecundity of limpets}
\format{An object of class \code{data.frame}}
\source{
Quinn & Keough (2000)
}
\usage{
data(snails)
}
\description{
Average number of eggs per snail, influenced by season
and population density \cr
More explanations in: Ex design, 04_Factorials_Part_II, p. 2 \cr
Example for: Two-way ANOVA, full factorial
}
\examples{
data(snails)
}
\keyword{datasets}
