% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigly.R
\docType{data}
\name{trigly}
\alias{trigly}
\title{Evaluating Machine Performance for triglyceride spectrophotometry}
\format{An object of class \code{data.frame}}
\source{
Kuehl (2000), Ex 7.1
}
\usage{
data(trigly)
}
\description{
Triglyceride level influenced by  machine and day \cr
More explanations in: Ex design, 06_Random_Effects, p. 30 \cr
Example for: Random effect Model
}
\examples{
data(trigly)
}
\keyword{datasets}
