% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique2010.R
\docType{data}
\name{unique2010}
\alias{unique2010}
\title{Data of Zurich airport}
\format{An object of class \code{data.frame}}
\usage{
data(unique2010)
}
\description{
ATM: Air traffic movement, Pax: number of passanger \cr
More explanations in: regression script p. 1 \cr
Example for: Simple regression
}
\examples{
data(unique2010)
}
\keyword{datasets}
